src/back_end/requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
autobahn==22.7.1
Automat==20.2.0
beautifulsoup4==4.11.1
bidict==0.22.0
billiard==3.6.4.0
bs4==0.0.1
celery==5.2.7
certifi==2022.6.15
cffi==1.15.1
channels==3.0.5
charset-normalizer==2.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.5
constantly==15.1.0
cryptography==37.0.4
cssbeautifier==1.14.5
cssselect==1.1.0
daphne==3.0.2
defusedxml==0.7.1
Deprecated==1.2.13
Django==4.0.7
django-appconf==1.0.5
django-celery-beat==2.3.0
django-celery-results==2.4.0
django-timezone-field==5.0
djangorestframework==3.13.1
EditorConfig==0.12.3
filelock==3.8.0
greenlet==1.1.3
gunicorn==20.1.0
html-tag-names==0.1.2
html-void-elements==0.1.0
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.12.0
incremental==21.3.0
itemadapter==0.7.0
itemloaders==1.0.4
jmespath==1.0.1
jsbeautifier==1.14.5
kombu==5.2.4
lxml==4.9.1
mccabe==0.7.0
packaging==21.3
parsel==1.6.0
pathspec==0.9.0
pilkit==2.0
Pillow==9.2.0
prompt-toolkit==3.0.30
Protego==0.2.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.9.1
pycparser==2.21
PyDispatcher==2.0.5
pyflakes==2.5.0
PyMySQL==1.0.2
pyOpenSSL==22.0.0
pyparsing==3.0.9
python-crontab==2.6.0
python-dateutil==2.8.2
python-scrapyd-api==2.1.2
python3-openid==3.2.0
pytz==2022.2.1
pywin32==304
queuelib==1.6.2
redis==4.3.4
regex==2022.7.25
requests==2.28.1
requests-file==1.5.1
Scrapy==2.6.2
scrapy-djangoitem==1.1.1
scrapyd==1.3.0
scrapyd-client==1.2.2
service-identity==21.1.0
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
tldextract==3.3.1
tomli==2.0.1
tqdm==4.64.0
Twisted==22.4.0
twisted-iocpsupport==1.0.2
txaio==22.2.1
typing_extensions==4.3.0
tzdata==2022.2
uberegg==0.1.1
urllib3==1.26.11
vine==5.0.0
w3lib==2.0.1
wcwidth==0.2.5
wrapt==1.14.1
yapf==0.32.0
zipp==3.8.1
zope.interface==5.4.0
